Refactor v2 parser pipe expressions
